app2/requirements.txt
flask
sqlalchemy
psycopg2-binary

gunicorn==20.1.0

opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-propagator-b3
opentelemetry.exporter.zipkin
opentelemetry-instrumentation-flask
opentelemetry.instrumentation.sqlalchemy
opentelemetry-instrumentation-logging